dashboard/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
altair>=5.0.0
numpy
matplotlib
seaborn
scipy
scikit-learn
umap-learn
minisom
plotly
